# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The basic auth token provided by distrybute and used to upload files.",
        "title": "distrybute API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/mmichaelb/distrybute/blob/master/LICENSE"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/file": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Upload a file using a POST request.",
                "operationId": "uploadFile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contains the file content which should be uploaded",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The response which contains the callReference",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.FileUploadResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/file/delete/{deleteReference}": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Deletes a specific file using the provided delete reference.",
                "operationId": "deleteFile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Call Reference",
                        "name": "deleteReference",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/v/{callReference}": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Retrieve a file by using the callReference parameter.",
                "operationId": "retrieveFile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Call Reference",
                        "name": "callReference",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.FileUploadResponse": {
            "type": "object",
            "properties": {
                "callReference": {
                    "type": "string"
                },
                "deleteReference": {
                    "type": "string"
                }
            }
        },
        "controller.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errorMessage": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}